rename: GetRequestHandler와 ResourceHandler의 Handler를 Adapter로 변경
